US-69: [madison] - add version endpoint, log in csum lambda
