+++ chalice/Makefile
@@ -3,8 +3,7 @@ include ../common.mk
 export EXPORT_ENV_VARS_TO_LAMBDA=DEPLOYMENT_STAGE \
 			INGEST_API_KEY \
 			INGEST_AMQP_SERVER \
-			API_HOST \
-			UPLOAD_SERVICE_VERSION
+			API_HOST
 
 clean:
 	git clean -df .
