+++ daemons/upload-checksum-daemon/Makefile
