+++ daemons/upload-checksum-daemon/requirements.txt
@@ -2,7 +2,7 @@ boto3==1.7.69
 crcmod==1.7
 pika==0.12.0
 requests==2.19.1
-dcplib==1.3.2
+dcplib==1.4.0
 SQLAlchemy==1.2.10
 alembic==1.0.0
 psycopg2-binary==2.7.5
