+++ requirements.txt
@@ -3,7 +3,7 @@ connexion==1.5.2
 crcmod==1.7
 pika==0.12.0
 requests==2.19.1
-dcplib==1.3.2
+dcplib==1.4.0
 proforma==0.2.1
 alembic==1.0.0
 psycopg2-binary==2.7.5
