+++ terraform/modules/upload-service/api_lambda.tf
@@ -115,8 +115,7 @@ resource "aws_lambda_function" "upload_api_lambda" {
       DEPLOYMENT_STAGE = "${var.deployment_stage}",
       INGEST_API_KEY = "${var.ingest_api_key}",
       INGEST_AMQP_SERVER = "${var.ingest_amqp_server}",
-      API_HOST = "${var.upload_api_fqdn}",
-      UPLOAD_SERVICE_VERSION = "${var.upload_service_version}"
+      API_HOST = "${var.upload_api_fqdn}"
     }
   }
   tags {
