+++ terraform/modules/upload-service/variables.tf
@@ -10,10 +10,6 @@ variable "deployment_stage" {
   type = "string"
 }
 
-variable "upload_service_version" {
-  type = "string"
-}
-
 // S3
 
 variable "bucket_name_prefix" {
