+++ upload/common/upload_config.py
@@ -5,3 +5,8 @@ class UploadConfig(Config):
 
     def __init__(self, *args, **kwargs):
         super().__init__('upload', **kwargs)
+
+
+class UploadVersion(Config):
+    def __init__(self, *args, **kwargs):
+        super().__init__(component_name='upload', secret_name='upload_service_version', **kwargs)
