+++ upload/lambdas/api_server/v1/area.py
@@ -1,5 +1,4 @@
 import json
-import os
 import urllib.parse
 import connexion
 import requests
@@ -13,13 +12,6 @@ from ....common.ingest_notifier import IngestNotifier
 from ....common.logging import get_logger
 
 logger = get_logger(__name__)
-
-UPLOAD_SERVICE_VERSION = os.getenv('UPLOAD_SERVICE_VERSION')
-
-
-@return_exceptions_as_http_errors
-def version():
-    return {'upload_service_version': UPLOAD_SERVICE_VERSION}, requests.codes.ok
 
 
 @return_exceptions_as_http_errors
